Numerous changes all stemming from work for the ability to select objects in screen space and have the renderer highlight them
- Implemented platform OS system for editor to allow EditorScene to receive os events
- Refactored editor window framework to be agnostic to the scene being run to support more than one editor scene type in the future
- Move editor logic into the editor scene rather than the general window object
- Added support for clients to receive mouse scroll wheel events
- Added ability to scroll to change global view scale in the editor
- Added ability to query for current mouse button state
- Support persistently mapped buffers and images
- Created separate Images subsystem in addition to Textures
- Refactored VMA usage to the latest recommend memory usage flags rather than deprecated flags
- Perf: Use dedicated memory allocations for render targets
- Refactored CMakeLists to intake a static version of PhysX
- Various tweaks to ensure projects build/run on Windows still
- Initial framework for client/renderer to query for entities/objects at a particular scene point
- Initial framework for adding a highlighting effect on/around objects
- Object renders now render an object+material output image
- Added render param for which objects to highlight
- Added render setting for which eye texture to present when in VR mode
- Translucent objects no longer cast shadows
- Added vcpkg baseline - newest SDL version has window startup validation error

+++ src/LibAccelaPlatform/include/Accela/Platform/Event/MouseMoveEvent.h
@@ -23,7 +23,7 @@ namespace Accela::Platform
          * @param xRel How far the mouse's x position changed, in pixels
          * @param yRel How far the mouse's y position changed, in pixels
          */
-        MouseMoveEvent(uint32_t _pointerId, unsigned int _xPos, unsigned int _yPos, int _xRel, int _yRel)
+        MouseMoveEvent(uint64_t _pointerId, float _xPos, float _yPos, float _xRel, float _yRel)
             : pointerId(_pointerId)
             , xPos(_xPos)
             , yPos(_yPos)
@@ -31,15 +31,15 @@ namespace Accela::Platform
             , yRel(_yRel)
         { }
 
-        uint32_t pointerId; // Unique id of the mouse, in multi-mouse setups
+        uint64_t pointerId; // Unique id of the mouse, in multi-mouse setups
 
         // These positions are in window space within the engine, and then overwritten with
         // virtual-space positions before being passed to the scene
-        unsigned int xPos;
-        unsigned int yPos;
+        float xPos;
+        float yPos;
 
-        int xRel;
-        int yRel;
+        float xRel;
+        float yRel;
     };
 }
